Move cloudChain.ts into src/chains
